Create analyzing-credit-card-spending-habits-in-india
The "Analyzing Credit Card Spending Habits in India" project is a part of the SQL Portfolio, showcasing an individual's expertise in SQL database management and data analysis. This project focuses on exploring and understanding the spending patterns and habits of credit card users in India through the analysis of transactional data.

Utilizing SQL queries and techniques, the project delves into a dataset containing credit card transactions from various banks and financial institutions in India. The objective is to extract meaningful insights and draw valuable conclusions about consumer behavior, preferences, and trends in credit card usage.